Refactor coach bot answer handling and streak management; introduce new utility files for better organization
